// src/app/models/car-models/car.ts
export class Car {
    id: string;
    brandName: string;
    isCommonBrand: boolean;
    modelName: string;
    modelAbbreviation: string;
    generationName: string;
    shortGenerationName: string;
    shortGenerationDescription: string;
    engineName: string;
    engineModelCode: string;
    power: string;
    powerPerLitre: string;
    torque: string;
    engineLayout: string;
    engineVolume: number;
    numberOfCylinders: number;
    engineConfiguration: string;
    cylinderBore: string;
    pistonStroke: string;
    compressionRatio: number;
    numberOfValvesPerCylinder: number;
    fuelInjectionSystem: string;
    engineAspiration: string;
    engineOilCapacity: number;
    coolant: number;
    engineSystems: string;
    engineOilSpecification: string;
    startOfProduction: string;
    endOfProduction: string;
    powertrainArchitecture: string;
    bodyType: string;
    seats: number;
    doors: number;
    fuelEconomyUrban: string;
    fuelEconomyExtraUrban: string;
    fuelEconomyCombined: string;
    acceleration0To100KmH: number;
    acceleration0To62MpH: number;
    fuelType: string;
    fuel: string;
    maximumSpeed: number;
    emissionStandard: string;
    weightToPowerRatio: string;
    weightToTorqueRatio: string;
    kerbWeight: number;
    maximumWeight: number;
    maximumLoad: number;
    trunkSpace: number;
    fuelTankCapacity: number;
    maximumRoofLoad: number;
    permittedTrailerLoadWithBrakes: number;
    permittedTrailerLoadWithoutBreaks: number;
    permittedTowBarDownload: number;
    length: number;
    width: number;
    height: number;
    wheelbase: number;
    frontTrack: number;
    rearTrack: number;
    frontOverhang: number;
    rearOverhang: number;
    rideHeight: number;
    dragCoefficient: number;
    turningDiameter: number;
    driveTrainArchitecture: string;
    driveWheel: string;
    typeOfGearbox: string;
    frontSuspension: string;
    rearSuspension: string;
    frontBrakes: string;
    rearBrakes:string;
    assistingSystems:string;
    steeringType: string;
    tiresSize: string;
    wheelRimsSize: string;
    powerSteering: string;
}